move evaluation tests to own file (#370)
* move evaluation tests to own file
